Category

+++ src/main/java/com/educaweb/course/entities/Category.java
@@ -24,7 +24,7 @@ public class Category implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 
 	private String name;
 
@@ -35,7 +35,7 @@ public class Category implements Serializable {
 	public Category() {
 	}
 
-	public Category(Integer id, String name) {
+	public Category(Long id, String name) {
 		super();
 		this.id = id;
 		this.name = name;
@@ -45,11 +45,11 @@ public class Category implements Serializable {
 		return products;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
